requirements.txt
streamlit==1.38.0
streamlit-pdf-viewer==0.0.18
pycryptodome==3.20.0